Fix handling of 'content' in model
The 'content' value is a rich object (hash or array), but the form wants
to deal with it as a string. Also, we need to be able to store invalid
input from the form so we can present it back to the user to get
corrections without reverting it to the old content.